requirements.txt
azure-cognitiveservices-speech==1.25.0
openai==0.26.4